+++ src/node/utils/ExportHelper.js
@@ -51,7 +51,7 @@ exports._analyzeLine = (text, aline, apool) => {
   let lineMarker = 0;
   line.listLevel = 0;
   if (aline) {
-    const opIter = Changeset.opIterator(aline);
+    const opIter = new Changeset.OpIter(aline);
     if (opIter.hasNext()) {
       const op = opIter.next();
       let listType = Changeset.opAttributeValue(op, 'list', apool);