+++ src/tests/frontend/specs/easysync.js
@@ -31,7 +31,7 @@ const randInt = (maxValue) => Math.floor(Math.random() * maxValue);
 describe('easysync', function () {
   it('throughIterator', async function () {
     const x = '-c*3*4+6|3=az*asdf0*1*2*3+1=1-1+1*0+1=1-1+1|c=c-1';
-    const iter = Changeset.opIterator(x);
+    const iter = new Changeset.OpIter(x);
     const assem = Changeset.opAssembler();
     while (iter.hasNext()) assem.append(iter.next());
     expect(assem.toString()).to.equal(x);
@@ -39,7 +39,7 @@ describe('easysync', function () {
 
   it('throughSmartAssembler', async function () {
     const x = '-c*3*4+6|3=az*asdf0*1*2*3+1=1-1+1*0+1=1-1+1|c=c-1';
-    const iter = Changeset.opIterator(x);
+    const iter = new Changeset.OpIter(x);
     const assem = Changeset.smartOpAssembler();
     while (iter.hasNext()) assem.append(iter.next());
     assem.endDocument();
@@ -730,7 +730,7 @@ describe('easysync', function () {
     p.putAttrib(['name', 'david']);
     p.putAttrib(['color', 'green']);
 
-    const stringOp = (str) => Changeset.opIterator(str).next();
+    const stringOp = (str) => new Changeset.OpIter(str).next();
 
     expect(Changeset.opAttributeValue(stringOp('*0*1+1'), 'name', p)).to.equal('david');
     expect(Changeset.opAttributeValue(stringOp('*0+1'), 'name', p)).to.equal('david');